Moved packing tests to integration.
